rubocop and fix exception issue, and move coerce_to_date
